edit questions

+++ api/view/theme.py
@@ -3,22 +3,26 @@ from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
 from api.models import Theme, Question
-from api.serializer import QuestionSerializer, ThemeSerializer
+from api.serializer import  ThemeDetailSerializer, ThemeListSerializer
 from api.view import is_authenticate
 
 
 @api_view(['GET'])
 def get_themes(request):
+    #get all theme without details
+
     user = is_authenticate(request)
     if not user:
         return Response({"message": "Erreur lors de l'authentification"})
 
     themes = Theme.objects.all()
-    serializer = ThemeSerializer(themes,many=True)
+    serializer = ThemeListSerializer(themes,many=True)
+
     return (Response(serializer.data))
 
 @api_view(['GET',"PUT","DELETE"])
 def get_theme(request,id):
+    #get one theme and details
     user = is_authenticate(request)
     if not user:
         return Response({"message": "Erreur lors de l'authentification"})
@@ -27,13 +31,13 @@ def get_theme(request,id):
     if theme.author == None:
         theme.delete()
     if request.method == 'GET':
-         return Response(ThemeSerializer(theme).data)
+         return Response(ThemeDetailSerializer(theme).data)
     if request.method == "PUT":
 
         if theme.author.id != user.id:
             return Response({"message": "You can't delete this"}, status=403)
 
-        serializer = ThemeSerializer(instance=theme,data=request.data)
+        serializer = ThemeDetailSerializer(instance=theme,data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return  Response(serializer.data)
@@ -41,10 +45,10 @@ def get_theme(request,id):
     if request.method == "DELETE":
         if theme.author.id != user.id:
             return Response({"message": "You can't delete this"}, status=403)
-
-        theme.delete()
-        return  Response({'message':"ok"})
-
+        if not theme.question_set.exists():
+             theme.delete()
+             return  Response({'message':"ok"})
+        return Response({"message": "Cannot delete a theme with associated questions"}, status=400)
 
 @api_view(['POST'])
 def create_theme(request):
@@ -58,7 +62,7 @@ def create_theme(request):
     data = request.data.copy()
     data['author'] = user.id
     print(data)
-    serializer = ThemeSerializer(data=data, partial=True)
+    serializer = ThemeDetailSerializer(data=data, partial=True)
     serializer.is_valid(raise_exception=True)
     question = serializer.save()
     print(question)
